+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Features
 * Removed `tea` module dependency for pre-existing types in `stdlib` module.
 * Support `file` as a `file_rolling_type`.
+* Added `java_opts` parameter to `elasticsearch::plugin` resource.
 
 #### Fixes