+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Removed `tea` module dependency for pre-existing types in `stdlib` module.
 * Support `file` as a `file_rolling_type`.
 * Added `java_opts` parameter to `elasticsearch::plugin` resource.
+* Brought some options in `jvm.options` up-to-date with upstream.
 
 #### Fixes