+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Added `java_opts` parameter to `elasticsearch::plugin` resource.
 
 #### Fixes
-* Fixed issue with `ES_PATH_CONF` being unset in CentOS 6 SysV init files.
+* Fixed issue with `ES_PATH_CONF` being unset in SysV init files.
 
 ## 6.0.0 (November 14, 2017)