+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 * Support `file` as a `file_rolling_type`.
 * Added `java_opts` parameter to `elasticsearch::plugin` resource.
 * Brought some options in `jvm.options` up-to-date with upstream.
+* Plugins can now have their `JAVA_HOME` set through the `java_home` parameter.
 
 #### Fixes
+* Fixed issue with `ES_PATH_CONF` being unset in SysV init files.
 
 ## 6.0.0 (November 14, 2017)